Refactor: Remove cardId from onPlay and onDeactivate effects
The cardId is now accessed via the LatestCardComponent on the player entity.
This change simplifies the function signatures for onPlay and onDeactivate effects.
A new LatestCardComponent has been added to the player entity to store the ID of the most recently played card.

Oh boy I wish there was tests for everything.
Or that I had energy and time to implement them...