Refactor Effect classes to use a common `amount` property
This commit refactors the various `Effect` data classes to inherit an `amount: Float` property from the abstract `Effect` class. This simplifies effect handling and standardizes the representation of effect magnitudes.

Additionally, a `getScoreF()` method was added to `ScoreComponent` to return the score as a float, and usages of `getScore()` were updated accordingly where a float was expected.